Allow pkimetal to listen on a Unix socket
We'd like to run pkimetal in parts of our CA infrastructure as a sidecar
to another process, and instead of worrying about networking, we would
prefer to just use a unix socket.

This adds three new configuration parameters:

 * server.webserverPath (no default)
 * server.monitoringPath (no default)
 * server.socketPermissions (default 0600)

If these are set, pkimetal will listen for requests on those unix sockets.
Listening in a port is now also conditional on the port being nonzero.
So with a default configuration, nothing changes.

I've added logging of the port/path listening to make it clearer what's happening.

The server.socketPermission sets the default permissions on the socket. 0600 is
a conservative default, and users can permit more open permissions.

I've been interactively testing this change like this:

   curl --unix-socket /tmp/pkimetal.sock http://x/lintcert \
        --data-urlencode b64input@cert.pem